+++ invokeai/backend/stable_diffusion/__init__.py
@@ -2,14 +2,8 @@
 Initialization file for the invokeai.backend.stable_diffusion package
 """
 from .diffusers_pipeline import (  # noqa: F401
-    ConditioningData,
     PipelineIntermediateState,
     StableDiffusionGeneratorPipeline,
 )
 from .diffusion import InvokeAIDiffuserComponent  # noqa: F401
 from .diffusion.cross_attention_map_saving import AttentionMapSaver  # noqa: F401
-from .diffusion.shared_invokeai_diffusion import (  # noqa: F401
-    PostprocessingSettings,
-    BasicConditioningInfo,
-    SDXLConditioningInfo,
-)
